Moving fitness model to model.dart
